chore: restructure blueprint files (#1006)
